requirements.txt
Django>=5.2.4
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-phonenumber-field>=7.1.0
phonenumbers>=8.13.0
pandas>=2.0.0
scikit-learn>=1.3.0
pdfplumber>=0.10.0
psycopg2-binary>=2.9.0

